Issue #254 Remove extended bit from path unknown path attribute
The unknown-attr model IETF inherited from the OpenConfig model included a
boolean flag for the "extended" bit in the Path Attribute flags.  This flag is
used to indicate that the length field after the Path Code is either one or two
octets in length.

This flag is not required to be preserved by the protocol since the length
itself is normally enough to indicate whether two-octet encoding is required.

There is an edge case that length < 256 octets for a path attribute may be
encoded as a two-octet field.  While this isn't the usual desired encoding,
it's still valid and introduces no headaches.  RFC 4271 and RFC 7606 are silent
about such a case.

Since implementations aren't required to preserve the received Extended flag,
remove it from this model and document the reason in the length field.