Issue #255 Address structure of secure-session
The enabled boolean didn't reside in the container with the secure-session
parameters themselves.  The constraint on the existing options had a constraint
that you couldn't configure options without setting the enabled status to true.
This is the opposite of how you'd normally want this to operate.

Move the enabled into the secure-session container.
Create a must constraint that when enabled, options MUST be configured.