Issue #277 - Hold and established time changes
Move the negotiated-hold-time leaf to the timers container.

Remove the fsm-established-time leaf.  The leaf, last-established, covers the
event when we transition in and out of Established.  The timer that's being
deleted can be derived from the last-established timestamp and the current
session-state.  This is also more consistent with the OpenConfig model.

Closes #277