Issue #253 - Remove redundant bfd container, change local-address
There was a redundant bfd container left after cleanup. It has been removed.
(There is still one covered in a different Issue #251.)

bgp/[...]/transport/local-address appears redundant, but in the current model
it is used to configure the local-address.  The
bgp/neighbors/transport/local-address is currently operational-only state.

Moved the local-address node up to the top of transport.