Issue #228 - A BGP YANG regular expression syntax
Per the issue, we need a format for our regular expressions.  Start with one
similar to that of RPSL.

The supplied patch is too lose from a formal language perspective to make
people happy long-term, but is enough to start IETF discussion.

An ABNF section will be added at a later date.