Issue #253 - Remove redundant bfd container, change local-address (#282)
* Issue #253 - Remove redundant bfd container, change local-address

There was a redundant bfd container left after cleanup. It has been removed.
(There is still one covered in a different Issue #251.)

bgp/[...]/transport/local-address appears redundant, but in the current model
it is used to configure the local-address.  The
bgp/neighbors/transport/local-address is currently operational-only state.

Moved the local-address node up to the top of transport.

* Address review comments, fix warning.

Mahesh wanted this moved to if:interface-name.
The compiler tells us that we didn't need the bfd-types import.

* Remove the bgp/interfaces container

This container existed for two reasons:
1. Provided a location for interface-refs for local-address in this diff.  Per
review comments, that now refers to the main IETF interface module.

2. Contained the per-interface BFD state.  According to Keyur, that's case was
from before the BFD YANG module addressed the use cases.

Closes #251

* Add reference to RFC 7223 in import to ietf-interfaces, per review

Co-authored-by: Jeff Haas <jhaas@juniper.net>