Issue #249 - Better operational state for errors (#289)
* Issue #249 - Better operational state for errors

The old style RFC 4273 MIB stuff wasn't great, even for MIBs.
Using inspiration from the OpenConfig identities, created a
new sub-module for the hierarchy of BGP notification codes and sub-codes.
Unlike the OC version, it's all anchored at a common root.  Most of the size of
this module is providing the normative references for what the identity
actually means.

Similarly, the model needs to carry the error-code and subcode for situations
where the identity doesn't provide sufficient information.

The notification data and time are also added to this state.

The YANG notifications for the established and backward-transition state are
similarly cleaned up and removed things that came directly from RFC 4273 that
weren't particularly helpful.  As part of that work, the backwards-transition
notification is now capable of putting directional data in it.

Closes #249
Closes #250

* Address some review comments

* Update the leaf discussing the identity to note that it will include best base error code.

* Change names of identities within notification hierarchy per discussion with Mahesh

* Update ietf-bgp.yang

Co-authored-by: Jeff Haas <jhaas@juniper.net>