Issue #281 Eliminate grouping neighbor-group-config
grouping neighbor-group-config is used in exactly the same places that grouping
neighbor-and-peer-group-common are used, so they're really the same thing.

Eliminate neighbor-group config.  This required restructuring some of the
dependencies of the submodules.

Note that this causes the graceful-restart config and state to bubble down in
the hierarchy where used in the neighbors and peer-groups.

Closes #281