Issue #309 Dynamic peers feature
Many implementations of BGP support dynamic peers.  Most tend to configure it
from peer-group scope.  This commit adds dynamic peers as a grouping that is
used at the per-group scope.  Its structure is a container with a single list
where the prefixes of the peers are configured.

While this modeling is mostly empty space, it enables two flavors of extensibility:
1. The list keyed by prefix can have additional per-prefix configuration.  This
   accommodates features in some vendors to further restrict peers on a
   per-prefix basis.
2. The empty container allows per-container augmentation as well.

If an implementation prefers to model this similar to OpenConfig with a
top-level bgp dynamic peers container, this grouping can be augmented in at the
global level to provide naming consistency.

Closes #309