Issue #225 A typedef for BGP color extended communities
I'm not sure we want to proceed with this commit in the initial IETF BGP YANG
module, but this is a useful modeling exercise for how to solve some ugly
issues.

The BGP Color community has changed in RFC 9012 from its older format to move
the first 16 bits of RESERVED value into a Flags field.  SR-TE policy work has
allocated two of those bits.  A number of bits are currently unknown and may be
assigned further in the future.

A flexible format where, when the flags are all known, the flag fields are
displayed as a comma-separated set of flag bits.  If any flag bits are unknown,
or the Flags are zero, a default format to render the Flags as a uint16 is
done.

Note that this varies from the OC proposal to simply display the two known
flags as it doesn't cover future flag possibilities.

Closes #225