Issue #306 Cleanup IANA BGP Types (#320)
* Issue #306 Cleanup IANA BGP Types

Sort the sections for the IANA BGP types and roughly sort the items within
them.  Since community-related stuff is expected to churn the most, these have
been pushed to the bottom of the file.

Closes #306

* BUG FIX: Update patterns to permit "0" to be part of the target numbers.

Add RFC 5668 4-octet specific route-origin and route-target.  These are used in
the real world.

Choose a different documentation format for the extended community types.
Ideally, in the context of a typedef, a type would permit a description and a
reference in them.  However, since these bits of the language aren't allowed,
instead pretend they're present and put them in a comment.  Perhaps Mahesh can
take the request for supporting this in YANG 1.2 to netmod.

* Missed some fixes to empty zero fix

* Change type/subtype to use hex format

Co-authored-by: Jeff Haas <jhaas@juniper.net>